clean up some battery stuff
