+++ src/canvas/drawing_utils.rs
@@ -1,4 +1,4 @@
-use std::{cmp::min, time::Instant};
+use std::time::Instant;
 
 use tui::{
     layout::Rect,
@@ -6,14 +6,6 @@ use tui::{
 };
 
 use super::SIDE_BORDERS;
-
-/// Calculate how many bars are to be drawn within basic mode's components.
-pub fn calculate_basic_use_bars(use_percentage: f64, num_bars_available: usize) -> usize {
-    min(
-        (num_bars_available as f64 * use_percentage / 100.0).round() as usize,
-        num_bars_available,
-    )
-}
 
 /// Determine whether a graph x-label should be hidden.
 pub fn should_hide_x_label(
@@ -63,20 +55,6 @@ pub fn dialog_block(border_type: BorderType) -> Block<'static> {
 mod test {
 
     use super::*;
-
-    #[test]
-    fn test_calculate_basic_use_bars() {
-        // Testing various breakpoints and edge cases.
-        assert_eq!(calculate_basic_use_bars(0.0, 15), 0);
-        assert_eq!(calculate_basic_use_bars(1.0, 15), 0);
-        assert_eq!(calculate_basic_use_bars(5.0, 15), 1);
-        assert_eq!(calculate_basic_use_bars(10.0, 15), 2);
-        assert_eq!(calculate_basic_use_bars(40.0, 15), 6);
-        assert_eq!(calculate_basic_use_bars(45.0, 15), 7);
-        assert_eq!(calculate_basic_use_bars(50.0, 15), 8);
-        assert_eq!(calculate_basic_use_bars(100.0, 15), 15);
-        assert_eq!(calculate_basic_use_bars(150.0, 15), 15);
-    }
 
     #[test]
     fn test_should_hide_x_label() {
