+++ src/canvas/widgets.rs
@@ -1,4 +1,3 @@
-pub mod battery_display;
 pub mod cpu_basic;
 pub mod cpu_graph;
 pub mod disk_table;
@@ -8,3 +7,6 @@ pub mod network_basic;
 pub mod network_graph;
 pub mod process_table;
 pub mod temperature_table;
+
+#[cfg(feature = "battery")]
+pub mod battery_display;
